+++ packages/plugin-workflow-approval/src/client/common/components/FormBlock.provider.tsx
@@ -14,8 +14,8 @@ import {
 } from '@tachybase/client';
 import { createForm, RecursionField, useField, useFieldSchema } from '@tachybase/schema';
 
-import { useContextApprovalExecution } from '../../user-interface/common/contexts/approvalExecution';
 import { useContextApprovalRecords } from '../../user-interface/pc/block/table-todos/providers/ApprovalExecutions.provider';
+import { useContextApprovalExecution } from '../contexts/approvalExecution';
 
 export const FormBlockProvider = (props) => {
   const { formType } = props;
